+++ public/static/index.html
@@ -33,8 +33,6 @@
       <input type="text" name="Full Name" />
       <input type="email" name="Email" />
       <input type="text" name="Organisation" />
-      <input name="Business Category" />
-      <input name="Region of Operations" />
       <textarea name="Message"></textarea>
       <input type="checkbox" name="Receive communications" />
     </form>
