+++ src/components/MainPageContainer.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { LandingSection } from "./HomePageContent/LandingSection";
 import { MainBenefitsSection } from "./HomePageContent/MainBenefitsSection";
 import { DocumentationSection } from "./HomePageContent/DocumentationSection";
-import { EmailSection } from "./HomePageContent/EmailSection";
 import { DropZoneSectionContainer } from "./HomePageContent/DropZoneSection";
 import { updateCertificate } from "../reducers/certificate";
 import { getLogger } from "../utils/logger";
@@ -29,7 +28,6 @@ const MainPage = ({ loadCertificate }: { loadCertificate: (certificate: any) => 
       <DropZoneSectionContainer />
       <LandingSection />
       <MainBenefitsSection />
-      <EmailSection />
       <DocumentationSection />
     </div>
   );
