+++ src/components/UI/Checkbox/Checkbox.tsx
@@ -1,34 +1,20 @@
-import styled from "@emotion/styled";
 import React, { InputHTMLAttributes } from "react";
 
 interface CheckboxProps extends InputHTMLAttributes<HTMLInputElement> {
   className?: string;
-  text: string;
+  children: React.ReactNode;
 }
 
-export const Checkbox = ({ className, text, ...props }: CheckboxProps) => {
+export const Checkbox = ({ className, children, ...props }: CheckboxProps) => {
   return (
     <label className={className}>
       <div className="flex items-center">
         <div className="w-auto">
           <input type="hidden" value="No" {...props} />
-          <input type="checkbox" value="Yes" {...props} />
+          <input className="mb-0 mr-2" type="checkbox" value="Yes" {...props} />
         </div>
-        <div className="flex-grow">
-          <p>{text}</p>
-        </div>
+        <div className="flex-grow">{children}</div>
       </div>
     </label>
   );
 };
-
-export const CheckboxDefault = styled(Checkbox)`
-  input,
-  p {
-    margin-bottom: 0;
-  }
-
-  input {
-    margin-right: 8px;
-  }
-`;
