+++ src/routes.tsx
@@ -6,6 +6,7 @@ import { PageNotFound } from "./pages/pageNotFound";
 import { ResourcesPage } from "./pages/resources";
 import { SettingsAddressBookPage, SettingsAddressResolverPage, SettingsPage } from "./pages/settings";
 import { ViewerPage } from "./pages/viewer";
+import { ContactPage } from "./pages/contact";
 
 export const routes = [
   { path: "/", exact: true, component: HomePageContainer },
@@ -17,5 +18,6 @@ export const routes = [
   { path: "/email/success", exact: true, component: EmailSuccessPage },
   { path: "/resources", exact: true, component: ResourcesPage },
   { path: "/media", exact: true, component: MediaPage },
+  { path: "/contact", exact: true, component: ContactPage },
   { path: "*", component: PageNotFound },
 ];
